+++ demisto_sdk/commands/doc_reviewer/tests/doc_reviewer_test.py
@@ -1,249 +1,271 @@
-import os
-from pathlib import Path
 from typing import List
 
 import pytest
 
 from demisto_sdk.commands.doc_reviewer.doc_reviewer import DocReviewer
-from demisto_sdk.tests.constants_test import GIT_ROOT
 from TestSuite.test_tools import ChangeCWD
 
 FIRST_KNOWN_WORDS_PATH = 'known_words_for_tests.txt'  # Contains: nomnomone, killaone
 SECOND_KNOWN_WORDS_PATH = 'another_known_words_for_tests.txt'  # Contains: nomnomtwo, killatwo
 
 
-class TestUsingKnownWordsFiles:
-    """Test class to setup the known words files."""
-    @staticmethod
-    def setup():
-        with open(FIRST_KNOWN_WORDS_PATH, 'w+') as file_to_check:
-            file_to_check.write("nomnomone\nkillaone")
+@pytest.mark.parametrize('file_content, unknown_words, known_words_files_contents, review_success',
+                         [("This is nomnomone, nomnomtwo", {},
+                           [["nomnomone", "killaone"], ["nomnomtwo", "killatwo"]], True),
+                          ("This is nomnomone, nomnomtwo", {"nomnomtwo": []},
+                           [["nomnomone", "killaone"]], False)])
+def test_having_two_known_words_files(repo, file_content, unknown_words, known_words_files_contents,
+                                      review_success):
+    """
+    Given:
+        - A release notes file with two misspelled words.
+        - Different variations of known_words files.
 
-        with open(SECOND_KNOWN_WORDS_PATH, 'w+') as file_to_check:
-            file_to_check.write("nomnomtwo\nkillatwo")
+    When:
+        - Running doc_reviewer with known_words_file_paths.
 
-    @staticmethod
-    def teardown():
-        if os.path.isfile(FIRST_KNOWN_WORDS_PATH):
-            os.remove(FIRST_KNOWN_WORDS_PATH)
+    Then:
+        - Ensure the review result is appropriate.
+        - Make sure a review has taken place.
+        - Enusure the unknown words are as expected.
+    """
+    pack = repo.create_pack('test_pack')
+    rn_file = pack.create_release_notes(version='1_0_0', content=file_content)
+    known_words_file_paths = []
+    for index, known_words_file_contents in enumerate(known_words_files_contents):
+        known_words_file = pack._create_text_based(f"known_words_{index}.txt")
+        known_words_file.write_list(known_words_file_contents)
+        known_words_file_paths.append(known_words_file.path)
 
-        if os.path.isfile(SECOND_KNOWN_WORDS_PATH):
-            os.remove(SECOND_KNOWN_WORDS_PATH)
-
-
-class TestSingleInputFile(TestUsingKnownWordsFiles):
-    """Test class for single input file"""
-    FILE_PATH_TO_CHECK = 'some_file.md'
-
-    def teardown(self):
-        super(TestSingleInputFile, self).teardown()
-        if os.path.isfile(self.FILE_PATH_TO_CHECK):
-            os.remove(self.FILE_PATH_TO_CHECK)
-
-    @pytest.mark.parametrize('file_content, unknown_words, known_words_file_paths, review_success',
-                             [("This is nomnomone, nomnomtwo", {},
-                               [FIRST_KNOWN_WORDS_PATH, SECOND_KNOWN_WORDS_PATH], True),
-                              ("This is nomnomone, nomnomtwo", {"nomnomtwo": []},
-                               [FIRST_KNOWN_WORDS_PATH], False)])
-    def test_having_two_known_words_files(self, mocker, file_content, unknown_words, known_words_file_paths,
-                                          review_success):
-        """
-        Given:
-            - A release notes file with two misspelled words.
-            - Different variations of known_words files.
-
-        When:
-            - Running doc_reviewer with known_words_file_paths.
-
-        Then:
-            - Ensure the review result is appropriate.
-            - Make sure a review has taken place.
-            - Enusure the unknown words are as expected.
-        """
-        with open(self.FILE_PATH_TO_CHECK, 'w+') as file_to_check:
-            file_to_check.write(file_content)
-
-        doc_reviewer = DocReviewer(file_paths=[self.FILE_PATH_TO_CHECK],
-                                   known_words_file_paths=known_words_file_paths)
-
-        release_notes_path = Path(f'SomePack/ReleaseNotes/{self.FILE_PATH_TO_CHECK}')
-        mocker.patch.object(Path, '__new__', return_value=release_notes_path)
-
+    with ChangeCWD(repo.path):
+        doc_reviewer = DocReviewer(file_paths=[rn_file.path], known_words_file_paths=known_words_file_paths)
         assert doc_reviewer.run_doc_review() == review_success
         assert len(doc_reviewer.files) > 0
         assert doc_reviewer.unknown_words == unknown_words
 
-    @pytest.mark.parametrize('file_content, unknown_words, known_words_file_paths, review_success',
-                             [("This is nomnomone, nomnomtwo", set(), [FIRST_KNOWN_WORDS_PATH], True),
-                              ("This is nomnomone, nomnomtwo", {"nomnomone"}, [], False)])
-    def test_adding_known_words_from_pack(self, mocker, file_content, unknown_words, known_words_file_paths,
-                                          review_success):
-        """
-        Given:
-            - A release notes file with two misspelled words.
-            - Different variations of known_words files.
 
-        When:
-            - Running doc_reviewer with known_words_file_paths and load_known_words_from_pack option.
+@pytest.mark.parametrize('file_content, unknown_words, known_words_files_contents, packs_known_words_content, '
+                         'review_success',
+                         [("This is nomnomone, nomnomtwo", set(), [["nomnomone"]], ["[known_words]", "nomnomtwo"], True),
+                          ("This is nomnomone, nomnomtwo", {"nomnomone"}, [], ["[known_words]", "nomnomtwo"], False),
+                          ("This is nomnomone, nomnomtwo, nomnomthree", {"nomnomthree"}, [["nomnomone"]],
+                           ["[known_words]", "nomnomtwo"], False),
+                          ("This is nomnomone, nomnomtwo, nomnomthree", set(),
+                           [["nomnomone"], ["nomnomthree"]], ["[known_words]", "nomnomtwo"], True)])
+def test_adding_known_words_from_pack(repo, file_content, unknown_words, known_words_files_contents,
+                                      packs_known_words_content, review_success):
+    """
+    Given:
+        - A release notes file with two misspelled words.
+        - Different variations of known_words files, including pack-ignore known_words.
 
-        Then:
-            - Ensure the review result is appropriate.
-            - Make sure a review has taken place.
-            - Enusure the unknown words are as expected.
-        """
-        with open(self.FILE_PATH_TO_CHECK, 'w+') as file_to_check:
-            file_to_check.write(file_content)
+    When:
+        - Running doc_reviewer with known_words_file_paths and load_known_words_from_pack option.
 
-        release_notes_path = Path(f'SomePack/ReleaseNotes/{self.FILE_PATH_TO_CHECK}')
-        mocker.patch.object(Path, '__new__', return_value=release_notes_path)
-        mocker.patch.object(DocReviewer, 'find_known_words_from_pack', return_value=SECOND_KNOWN_WORDS_PATH)
+    Then:
+        - Ensure the review result is appropriate.
+        - Make sure a review has taken place.
+        - Enusure the unknown words are as expected.
+    """
+    pack = repo.create_pack('test_pack')
+    rn_file = pack.create_release_notes(version='1_0_0', content=file_content)
+    pack.pack_ignore.write_list(packs_known_words_content)
+    known_words_file_paths = []
+    for index, known_words_file_contents in enumerate(known_words_files_contents):
+        known_words_file = pack._create_text_based(f"known_words_{index}.txt")
+        known_words_file.write_list(known_words_file_contents)
+        known_words_file_paths.append(known_words_file.path)
 
-        doc_reviewer = DocReviewer(file_paths=[self.FILE_PATH_TO_CHECK],
+    with ChangeCWD(repo.path):
+        doc_reviewer = DocReviewer(file_paths=[rn_file.path],
                                    known_words_file_paths=known_words_file_paths,
                                    load_known_words_from_pack=True)
-
         assert doc_reviewer.run_doc_review() == review_success
         assert len(doc_reviewer.files) > 0
         assert set(doc_reviewer.unknown_words.keys()) == unknown_words
 
 
-class TestTwoInputFiles(TestUsingKnownWordsFiles):
-    FIRST_FILE_PATH = 'some_file.md'
-    SECOND_FILE_PATH = 'some_other_file.md'
-
-    def teardown(self):
-        super(TestTwoInputFiles, self).teardown()
-        if os.path.isfile(self.FIRST_FILE_PATH):
-            os.remove(self.FIRST_FILE_PATH)
-
-        if os.path.isfile(self.SECOND_FILE_PATH):
-            os.remove(self.SECOND_FILE_PATH)
-
-    @pytest.mark.parametrize('first_path, second_path, first_file_content, second_file_content, unknown_word_calls, '
-                             'known_words_file_paths, review_success, misspelled_files_num',
-                             [('SomePack/ReleaseNotes', 'SomePack/ReleaseNotes',  # Same pack
-                               "This is nomnomone, nomnomtwo", "This is killa", [],
-                               [FIRST_KNOWN_WORDS_PATH, SECOND_KNOWN_WORDS_PATH], True, 0),
-                              ('SomePack/ReleaseNotes', 'SomePack/ReleaseNotes',  # Same pack
-                               "This is nomnomone, nomnomtwo", "This is killa", [{"nomnomtwo": []}],
-                               [FIRST_KNOWN_WORDS_PATH], False, 1),
-                              ('SomePack/ReleaseNotes', 'SomePack/ReleaseNotes',  # Same pack
-                               "This is nomnomone, nomnomtwo", "This is killa, killatwo", [{"killatwo": []},
-                                                                                           {"nomnomtwo": []}],
-                               [FIRST_KNOWN_WORDS_PATH], False, 2),
-                              ('SomePack/ReleaseNotes', 'SomeOtherPack/ReleaseNotes',  # not same pack
-                               "This is nomnomone, nomnomtwo", "This is killa", [],
-                               [FIRST_KNOWN_WORDS_PATH, SECOND_KNOWN_WORDS_PATH], True, 0),
-                              ('SomePack/ReleaseNotes', 'SomeOtherPack/ReleaseNotes',  # not same pack
-                               "This is nomnomone, nomnomtwo", "This is killa", [{"nomnomtwo": []}],
-                               [FIRST_KNOWN_WORDS_PATH], False, 1),
-                              ('SomePack/ReleaseNotes', 'SomeOtherPack/ReleaseNotes',  # not same pack
-                               "This is nomnomone, nomnomtwo", "This is killa, killatwo", [{"killatwo": []},
-                                                                                           {"nomnomtwo": []}],
-                               [FIRST_KNOWN_WORDS_PATH], False, 2)
-                              ])
-    def test_having_two_file_paths(self, mocker, first_path, second_path, first_file_content,
-                                   second_file_content, unknown_word_calls, known_words_file_paths, review_success,
-                                   misspelled_files_num):
-        """
-        Given:
-            - 2 release notes files with two misspelled words each.
-            - Different variations of known_words files.
-
-        When:
-            - Running doc_reviewer with known_words_file_paths.
-
-        Then:
-            - Ensure the review result is appropriate.
-            - Make sure a review has taken place.
-            - Enusure the unknown words are as expected for each file.
-        """
-        with open(self.FIRST_FILE_PATH, 'w+') as file_to_check:
-            file_to_check.write(first_file_content)
-
-        with open(self.SECOND_FILE_PATH, 'w+') as file_to_check:
-            file_to_check.write(second_file_content)
-
-        unknown_word_calls_with_mocker = []
-        for unknown_words in unknown_word_calls:
-            unknown_word_calls_with_mocker.append(mocker.call(unknown_words=unknown_words))
-
-        first_release_notes_path = Path(f'{first_path}/{self.FIRST_FILE_PATH}')
-        second_release_notes_path = Path(f'{second_path}/{self.SECOND_FILE_PATH}')
-        mocker.patch.object(Path, '__new__', side_effect=[first_release_notes_path, second_release_notes_path])
-        print_unknown_words = mocker.patch.object(DocReviewer, 'print_unknown_words')
-
-        doc_reviewer = DocReviewer(file_paths=[self.FIRST_FILE_PATH, self.SECOND_FILE_PATH],
-                                   known_words_file_paths=known_words_file_paths)
-
-        assert doc_reviewer.run_doc_review() == review_success
-        assert len(doc_reviewer.files) == 2
-        print_unknown_words.assert_has_calls(unknown_word_calls_with_mocker, any_order=True)
-        assert len(doc_reviewer.files_with_misspells) == misspelled_files_num
-
-    @pytest.mark.parametrize('first_path, second_path, first_file_content, second_file_content, unknown_word_calls, '
-                             'known_words_file_paths, packs_unknown_words, review_success, misspelled_files_num',
-                             [('SomePack/ReleaseNotes', 'SomePack/ReleaseNotes',  # same pack
-                               "This is nomnomone, nomnomtwo", "This is killaone, killatwo", [],
-                               [FIRST_KNOWN_WORDS_PATH], [SECOND_KNOWN_WORDS_PATH, SECOND_KNOWN_WORDS_PATH], True, 0),
-                              ('SomePack/ReleaseNotes', 'SomeOtherPack/ReleaseNotes',  # not same pack
-                               "This is nomnomone, nomnomtwo", "This is killaone, killatwo",
-                               [{"nomnomtwo": []}, {"killaone": []}],
-                                 [], [FIRST_KNOWN_WORDS_PATH, SECOND_KNOWN_WORDS_PATH], False, 2)
-                              ])
-    def test_two_files_adding_known_words_from_pack(self, mocker, first_path, second_path, first_file_content,
-                                                    second_file_content, unknown_word_calls, known_words_file_paths,
-                                                    packs_unknown_words, review_success, misspelled_files_num):
-        """
-        Given:
-            - 2 release notes files with two misspelled words each.
-            - Different variations of known_words files.
-
-        When:
-            - Running doc_reviewer with known_words_file_paths and load_known_words_from_pack option.
-
-        Then:
-            - Ensure the review result is appropriate.
-            - Make sure a review has taken place.
-            - Enusure the unknown words are as expected for each file.
-        """
-        with open(self.FIRST_FILE_PATH, 'w+') as file_to_check:
-            file_to_check.write(first_file_content)
-
-        with open(self.SECOND_FILE_PATH, 'w+') as file_to_check:
-            file_to_check.write(second_file_content)
-
-        unknown_word_calls_with_mocker = []
-        for unknown_words in unknown_word_calls:
-            unknown_word_calls_with_mocker.append(mocker.call(unknown_words=unknown_words))
-
-        first_release_notes_path = Path(f'{first_path}/{self.FIRST_FILE_PATH}')
-        second_release_notes_path = Path(f'{second_path}/{self.SECOND_FILE_PATH}')
-        mocker.patch.object(Path, '__new__', side_effect=[first_release_notes_path, second_release_notes_path])
-        print_unknown_words = mocker.patch.object(DocReviewer, 'print_unknown_words', side_effect=DocReviewer.print_unknown_words)
-        print(f"PACKS known words {packs_unknown_words}")
-        mocker.patch.object(DocReviewer, 'find_known_words_from_pack', side_effect=packs_unknown_words)
-
-        doc_reviewer = DocReviewer(file_paths=[self.FIRST_FILE_PATH, self.SECOND_FILE_PATH],
-                                   known_words_file_paths=known_words_file_paths,
-                                   load_known_words_from_pack=True)
-
-        assert doc_reviewer.run_doc_review() == review_success
-        assert len(doc_reviewer.files) == 2
-        print_unknown_words.assert_has_calls(unknown_word_calls_with_mocker, any_order=True)
-        assert len(doc_reviewer.files_with_misspells) == misspelled_files_num
-
-
-@pytest.mark.parametrize('repo_path, file_path, expected_result',
-                         [(f"{GIT_ROOT}/demisto_sdk/tests/test_files", "Packs/DummyPack/ReleaseNotes/1_0_1.md",
-                           "Packs/DummyPack/known_words.txt"),
-                          (f"{GIT_ROOT}/demisto_sdk/tests/test_files", "Packs/CortexXDR/ReleaseNotes/1_1_1.md",
-                           ""),  # No known_words file there.
-                          (".", "some_fake_path", "")])
-def test_find_known_words_from_pack(repo_path, file_path, expected_result):
+@pytest.mark.parametrize('first_file_content, second_file_content, unknown_word_calls, known_words_files_contents, '
+                         'review_success, misspelled_files_num, packs_known_words_content, load_known_words_from_pack',
+                         [("This is nomnomone, nomnomtwo", "This is killa", [],
+                           [["nomnomone", "killaone"], ["nomnomtwo", "killatwo"]], True, 0, [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killa", [{"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 1, [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killa, killatwo", [{"killatwo": []},
+                                                                                       {"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 2, [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killa", [],
+                           [["nomnomone", "killaone"]], True, 0, ["[known_words]", "nomnomtwo", "killatwo"], True)
+                          ])
+def test_having_two_file_paths_same_pack(repo, mocker, first_file_content, second_file_content, unknown_word_calls,
+                                         known_words_files_contents, review_success, misspelled_files_num,
+                                         packs_known_words_content, load_known_words_from_pack):
     """
     Given:
-        - Different file paths.
+        - 2 release notes files with two misspelled words each.
+        - Different variations of known_words files, including pack-ignore known_words.
+
+    When:
+        - Running doc_reviewer with known_words_file_paths.
+
+    Then:
+        - Ensure the review result is appropriate.
+        - Make sure a review has taken place.
+        - Enusure the unknown words are as expected for each file.
+    """
+    pack = repo.create_pack('first_test_pack')
+    first_rn_file = pack.create_release_notes(version='1_0_0', content=first_file_content)
+    second_rn_file = pack.create_release_notes(version='1_0_1', content=second_file_content)
+    pack.pack_ignore.write_list(packs_known_words_content)
+    known_words_file_paths = []
+    for index, known_words_file_contents in enumerate(known_words_files_contents):
+        known_words_file = pack._create_text_based(f"known_words_{index}.txt")
+        known_words_file.write_list(known_words_file_contents)
+        known_words_file_paths.append(known_words_file.path)
+
+    unknown_word_calls_with_mocker = []
+    for unknown_words in unknown_word_calls:
+        unknown_word_calls_with_mocker.append(mocker.call(unknown_words=unknown_words))
+
+    print_unknown_words = mocker.patch.object(DocReviewer, 'print_unknown_words')
+
+    with ChangeCWD(repo.path):
+        doc_reviewer = DocReviewer(file_paths=[first_rn_file.path, second_rn_file.path],
+                                   known_words_file_paths=known_words_file_paths,
+                                   load_known_words_from_pack=load_known_words_from_pack)
+        assert doc_reviewer.run_doc_review() == review_success
+        assert len(doc_reviewer.files) == 2
+        print_unknown_words.assert_has_calls(unknown_word_calls_with_mocker, any_order=True)
+        assert len(doc_reviewer.files_with_misspells) == misspelled_files_num
+
+
+@pytest.mark.parametrize('first_file_content, second_file_content, unknown_word_calls, known_words_files_contents, '
+                         'review_success, misspelled_files_num, first_packs_known_words_content, '
+                         'second_packs_known_words_content, load_known_words_from_pack',
+                         [("This is nomnomone, nomnomtwo", "This is killaone", [],
+                           [["nomnomone", "killaone"], ["nomnomtwo", "killatwo"]], True, 0, [], [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killaone", [{"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 1, [], [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killaone, killatwo", [{"killatwo": []},
+                                                                                          {"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 2, [], [], False),
+
+                          ("This is nomnomone, nomnomtwo", "This is killaone, killatwo", [{"nomnomtwo": []},
+                                                                                          {"killaone": []}],
+                           [], False, 2, ["[known_words]", "nomnomone", "killaone"],
+                           ["[known_words]", "nomnomtwo", "killatwo"], True),
+
+                          ("This is killaone, nomnomone", "This is killatwo, nomnomtwo", [],
+                           [], True, 0, ["[known_words]", "nomnomone", "killaone"],
+                           ["[known_words]", "nomnomtwo", "killatwo"], True),
+                          ])
+def test_having_two_file_paths_different_pack(repo, mocker, first_file_content, second_file_content, unknown_word_calls,
+                                              known_words_files_contents, review_success, misspelled_files_num,
+                                              first_packs_known_words_content, second_packs_known_words_content, load_known_words_from_pack):
+    """
+    Given:
+        - 2 release notes files with two misspelled words each.
+        - Different variations of known_words files, including pack-ignore known_words.
+
+    When:
+        - Running doc_reviewer with known_words_file_paths.
+
+    Then:
+        - Ensure the review result is appropriate.
+        - Make sure a review has taken place.
+        - Enusure the unknown words are as expected for each file.
+    """
+    first_pack = repo.create_pack('first_test_pack')
+    second_pack = repo.create_pack('second_test_pack')
+    first_rn_file = first_pack.create_release_notes(version='1_0_0', content=first_file_content)
+    second_rn_file = second_pack.create_release_notes(version='1_0_1', content=second_file_content)
+    first_pack.pack_ignore.write_list(first_packs_known_words_content)
+    second_pack.pack_ignore.write_list(second_packs_known_words_content)
+    known_words_file_paths = []
+    for index, known_words_file_contents in enumerate(known_words_files_contents):
+        known_words_file = first_pack._create_text_based(f"known_words_{index}.txt")
+        known_words_file.write_list(known_words_file_contents)
+        known_words_file_paths.append(known_words_file.path)
+
+    unknown_word_calls_with_mocker = []
+    for unknown_words in unknown_word_calls:
+        unknown_word_calls_with_mocker.append(mocker.call(unknown_words=unknown_words))
+
+    print_unknown_words = mocker.patch.object(DocReviewer, 'print_unknown_words')
+
+    with ChangeCWD(repo.path):
+        doc_reviewer = DocReviewer(file_paths=[first_rn_file.path, second_rn_file.path],
+                                   known_words_file_paths=known_words_file_paths,
+                                   load_known_words_from_pack=load_known_words_from_pack)
+        assert doc_reviewer.run_doc_review() == review_success
+        assert len(doc_reviewer.files) == 2
+        print_unknown_words.assert_has_calls(unknown_word_calls_with_mocker, any_order=True)
+        assert len(doc_reviewer.files_with_misspells) == misspelled_files_num
+
+
+@pytest.mark.parametrize('first_file_content, second_file_content, unknown_word_calls, known_words_files_contents, '
+                         'review_success, misspelled_files_num, packs_known_words_content, load_known_words_from_pack',
+                         [("This is nomnomone, nomnomtwo", "This is killa", [],
+                           [["nomnomone", "killaone"], ["nomnomtwo", "killatwo"]], True, 0, [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killa", [{"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 1, [], False),
+                          ("This is nomnomone, nomnomtwo", "This is killa, killatwo", [{"killatwo": []},
+                                                                                       {"nomnomtwo": []}],
+                           [["nomnomone", "killaone"]], False, 2, [], False),
+                          ])
+def test_having_two_file_paths_not_same_pack(repo, mocker, first_file_content, second_file_content, unknown_word_calls,
+                                             known_words_files_contents, review_success, misspelled_files_num,
+                                             packs_known_words_content, load_known_words_from_pack):
+    """
+    Given:
+        - 2 release notes files with two misspelled words each.
+        - Different variations of known_words files, including pack-ignore known_words.
+
+    When:
+        - Running doc_reviewer with known_words_file_paths.
+
+    Then:
+        - Ensure the review result is appropriate.
+        - Make sure a review has taken place.
+        - Enusure the unknown words are as expected for each file.
+    """
+    pack = repo.create_pack('first_test_pack')
+    first_rn_file = pack.create_release_notes(version='1_0_0', content=first_file_content)
+    second_rn_file = pack.create_release_notes(version='1_0_1', content=second_file_content)
+    pack.pack_ignore.write_list(packs_known_words_content)
+    known_words_file_paths = []
+    for index, known_words_file_contents in enumerate(known_words_files_contents):
+        known_words_file = pack._create_text_based(f"known_words_{index}.txt")
+        known_words_file.write_list(known_words_file_contents)
+        known_words_file_paths.append(known_words_file.path)
+
+    unknown_word_calls_with_mocker = []
+    for unknown_words in unknown_word_calls:
+        unknown_word_calls_with_mocker.append(mocker.call(unknown_words=unknown_words))
+
+    print_unknown_words = mocker.patch.object(DocReviewer, 'print_unknown_words')
+
+    with ChangeCWD(repo.path):
+        doc_reviewer = DocReviewer(file_paths=[first_rn_file.path, second_rn_file.path],
+                                   known_words_file_paths=known_words_file_paths,
+                                   load_known_words_from_pack=load_known_words_from_pack)
+        assert doc_reviewer.run_doc_review() == review_success
+        assert len(doc_reviewer.files) == 2
+        print_unknown_words.assert_has_calls(unknown_word_calls_with_mocker, any_order=True)
+        assert len(doc_reviewer.files_with_misspells) == misspelled_files_num
+
+
+@pytest.mark.parametrize('known_words_content, expected_known_words',
+                         [(['[known_words]', 'wordament'], ['wordament']),
+                          (['[known_words]'], []),
+                          ([], [])])
+def test_find_known_words_from_pack(repo, known_words_content, expected_known_words):
+    """
+    Given:
+        - Pack's structure is correct and pack-ignore file is present.
+            - Case A: pack-ignore file has known_words section with words.
+            - Case B: pack-ignore file has known_words section without words.
+            - Case C: pack-ignore file doesn't have a known_words section.
 
     When:
         - Running DocReviewer.find_known_words_from_pack.
@@ -251,9 +273,13 @@ def test_find_known_words_from_pack(repo_path, file_path, expected_result):
     Then:
         - Ensure the found path result is appropriate.
     """
-    with ChangeCWD(repo_path):
-        doc_reviewer = DocReviewer(file_paths=[])
-        assert doc_reviewer.find_known_words_from_pack(file_path) == expected_result
+    pack = repo.create_pack('test_pack')
+    rn_file = pack.create_release_notes(version='1_0_0', content='Some release note')
+    pack.pack_ignore.write_list(known_words_content)
+    doc_reviewer = DocReviewer(file_paths=[])
+    with ChangeCWD(repo.path):
+        assert doc_reviewer.find_known_words_from_pack(rn_file.path) == ('Packs/test_pack/.pack-ignore',
+                                                                         expected_known_words)
 
 
 def test_camel_case_split():